Changing value of Strategies to double

+++ Java/HungrySkynet.java
@@ -4,18 +4,18 @@ import java.util.PriorityQueue;
 public class HungrySkynet extends Skynet {
 
   public interface Strategy {
-    public int value();
+    public double value();
     public Path getPath();
-    public int evaluate(BoardState currentBoard);
+    public double evaluate(BoardState currentBoard);
   }
 
   public class Shave implements Strategy
   {
-    int value = -1;
+    double value = -1;
     Path stratPath = new Path();
     BoardState state;
     @Override
-    public int value() {
+    public double value() {
       return value;
     }
 
@@ -25,7 +25,7 @@ public class HungrySkynet extends Skynet {
     }
 
     @Override
-    public int evaluate(BoardState currentBoard) {
+    public double evaluate(BoardState currentBoard) {
       state = currentBoard;
       
       state.board.tick(Robot.Move.Shave);
@@ -41,8 +41,8 @@ public class HungrySkynet extends Skynet {
 
       @Override
       public int compare(Strategy arg0, Strategy arg1) {
-        return (arg0.value() - arg1.value() == 0 ? 0 : arg0.value()
-            - arg1.value());
+        return (arg0.value() - arg1.value() == 0 ? 0 : (int)(arg0.value()
+                                                             - arg1.value()));
       }
     });
   }
